make new basin_query_mpi itno basin_query_mpi_complete
