+++ src/basin/Makefile.am
@@ -4,6 +4,7 @@ bin_PROGRAMS = basin_query
 
 if UCVM_HAVE_MPI
 bin_PROGRAMS += basin_query_mpi
+bin_PROGRAMS += basin_query_mpi_complete
 endif
 
 # General compiler/linker flags
@@ -19,6 +20,7 @@ basin_query_SOURCES = basin_query.c
 
 if UCVM_HAVE_MPI
 basin_query_mpi_SOURCES = basin_query_mpi.c
+basin_query_mpi_complete_SOURCES = basin_query_mpi_complete.c
 endif
 
 all: $(bin_PROGRAMS)
@@ -32,6 +34,9 @@ basin_query: basin_query.o
 	$(CC) -o $@ $^ $(AM_LDFLAGS)
 
 basin_query_mpi: basin_query_mpi.o
+	$(CC) -o $@ $^ $(AM_LDFLAGS)
+
+basin_query_mpi_complete: basin_query_mpi_complete.o
 	$(CC) -o $@ $^ $(AM_LDFLAGS)
 
 ############################################
